# openapi.yaml
openapi: 3.0.3
info:
  title: RHESS — Red Hat Enterprise Skills Server
  description: >-
    Self-hosted AI agent skills directory. Skills Catalog and Discovery endpoints are unauthenticated. Source management
    (write) endpoints require `Authorization: Bearer <RHESS_ADMIN_TOKEN>`.
  version: 0.1.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Value of the RHESS_ADMIN_TOKEN environment variable
...
paths:
...
  /api/v1/skills/search:
    get:
      summary: Fuzzy search skills
      tags:
        - Skills
      description: Full-text fuzzy search over skill names, descriptions, and source identifiers.
      parameters:
        - schema:
            type: string
          in: query
          name: q
          required: true
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: Internal numeric skill ID
                        source:
                          type: string
                          description: Source slug
                        sourceLabel:
                          type: string
                          description: Human-readable source label
                        sourceUrl:
                          type: string
                          nullable: true
                          description: Source git repository URL
                        slug:
                          type: string
                          description: Skill slug within its source
                        name:
                          type: string
                          description: Skill display name from frontmatter
                        description:
                          type: string
                          description: Skill description from frontmatter
                        artifactType:
                          type: string
                          enum:
                            - skill-md
                            - archive
                          description: Artifact type
                        digest:
                          type: string
                          description: SHA-256 content digest prefixed with 'sha256:'
                        category:
                          type: string
                          nullable: true
                          description: Skill category (null until spec support is added)
                        allowedTools:
                          type: array
                          items:
                            type: string
                          description: Tools the skill is allowed to use
                        skillPath:
                          type: string
                          description: Relative path to SKILL.md within the source repository
                        frontmatter:
                          type: object
                          additionalProperties: true
                          description: Full parsed frontmatter map (excluding name/description)
                        installCommand:
                          type: string
                          description: npx skills add … command to install this skill
                        lastModified:
                          type: string
                          description: ISO 8601 timestamp of last update
                      required:
                        - id
                        - source
                        - slug
                        - name
                        - description
                        - artifactType
                        - digest
                        - allowedTools
                        - skillPath
                        - installCommand
                        - lastModified
                  total:
                    type: integer
                    description: Total number of matching skills
                  query:
                    type: string
                    description: The search query that was executed
                required:
                  - data
                  - total
                  - query
        '400':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
  /api/v1/skills/:
    get:
      summary: List all skills
      tags:
        - Skills
      description: Returns a paginated list of all indexed skills.
      parameters:
        - schema:
            type: integer
            minimum: 1
            default: 1
          in: query
          name: page
          required: false
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          in: query
          name: per_page
          required: false
        - schema:
            type: string
            enum:
              - name
              - updated_at
            default: name
          in: query
          name: sort
          required: false
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: Internal numeric skill ID
                        source:
                          type: string
                          description: Source slug
                        sourceLabel:
                          type: string
                          description: Human-readable source label
                        sourceUrl:
                          type: string
                          nullable: true
                          description: Source git repository URL
                        slug:
                          type: string
                          description: Skill slug within its source
                        name:
                          type: string
                          description: Skill display name from frontmatter
                        description:
                          type: string
                          description: Skill description from frontmatter
                        artifactType:
                          type: string
                          enum:
                            - skill-md
                            - archive
                          description: Artifact type
                        digest:
                          type: string
                          description: SHA-256 content digest prefixed with 'sha256:'
                        category:
                          type: string
                          nullable: true
                          description: Skill category (null until spec support is added)
                        allowedTools:
                          type: array
                          items:
                            type: string
                          description: Tools the skill is allowed to use
                        skillPath:
                          type: string
                          description: Relative path to SKILL.md within the source repository
                        frontmatter:
                          type: object
                          additionalProperties: true
                          description: Full parsed frontmatter map (excluding name/description)
                        installCommand:
                          type: string
                          description: npx skills add … command to install this skill
                        lastModified:
                          type: string
                          description: ISO 8601 timestamp of last update
                      required:
                        - id
                        - source
                        - slug
                        - name
                        - description
                        - artifactType
                        - digest
                        - allowedTools
                        - skillPath
                        - installCommand
                        - lastModified
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                        description: Total number of indexed skills
                      page:
                        type: integer
                        description: Current page number (1-based)
                      per_page:
                        type: integer
                        description: Results per page
                      total_pages:
                        type: integer
                        description: Total number of pages
                      sort:
                        type: string
                        description: Active sort field
                    required:
                      - total
                      - page
                      - per_page
                      - total_pages
                      - sort
                required:
                  - data
                  - meta
        '400':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
  /api/v1/skills/{source}/{slug}/artifact:
    get:
      summary: Download skill artifact
      tags:
        - Discovery
      description: >-
        Returns the raw artifact: `text/markdown` for single-file skills, `application/gzip` (tar.gz) for multi-file
        skills.
      parameters:
        - schema:
            type: string
          in: path
          name: source
          required: true
        - schema:
            type: string
          in: path
          name: slug
          required: true
      responses:
        '200':
          description: Raw artifact.
          content:
            text/markdown:
              schema:
                type: string
            application/gzip:
              schema:
                type: string
                format: binary
        '404':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
...